+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Change log
 .. _Next version: https://github.com/matthiask/speckenv/compare/3.1...main
 
 - Added support for ``dummy://`` cache URLs.
+- Added support for parsing email URLs.
 
 
 `3.1`_ (2022-02-07)